requirements.txt
click==8.0.3
numpy==1.24.3
python-dotenv==0.19.2
requests==2.26.0
dnspython==2.2.1
et-xmlfile==1.1.0
requests==2.26.0
xlsxwriter==3.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
openpyxl==3.0.9
pyodbc==4.0.34
pyzbar==0.1.9
greenlet==1.1.3.post0
eventlet==0.33.1
werkzeug==2.2.2
lxml==4.9.1
huggingface-hub==0.14.1
datasets==2.11.0
transformers==4.28.1
setfit==0.7.0
fastapi==0.95.1
uvicorn==0.20.0
clean-text==0.6.0
